Add focus sidebar and focus group/pane shortcuts
Questions about the keyboard shortcuts for focusing on the sidebar or one of the groups/panes are very common online, as this is a very basic use case that missing from many cheatsheets and guides. I'd love to see it added here, as I believe it will help many Dash+Sublime Text 3 users.