Update Gitmoji cheatsheet.
I've updated the cheatsheet with the exact data from the gitmoji project.
There are some additions, deletions and updates, so I've just took the json and transformed it to the cheatsheet format.

To simplify this process, and the future updates, I've made the choice to remove the categories, and keep the original order. With that, we can totally automatize the update process.

Thanks for the work!