Add visidata cheatsheet.
I wasn't sure the best way to do this since the cheatsheet has already been made,
but needed to be ported to Dash's format. I chose to just paste the entire source
YAML file in, as I thought it would be easier to update in the future if/when
jsvine updates his documentation. I'm open to other ideas, though.